+++ console/cypress/integration/data/functions/test.ts
@@ -8,6 +8,8 @@ import {
   trackFunction,
   verifyPermissionTab,
   testSessVariable,
+  trackVolatileFunction,
+  trackVolatileFunctionAsQuery,
 } from './spec';
 import { getIndexRoute } from '../../../helpers/dataHelpers';
 
@@ -31,6 +33,8 @@ export const runCreateCustomFunctionsTableTests = () => {
     it('Verify permission tab', verifyPermissionTab);
     it('Delete custom function', deleteCustomFunction);
     it('Test custom function with Session Argument', testSessVariable);
+    it('Tracks VOLATILE function as mutation', trackVolatileFunction);
+    it('Tracks VOLATILE function as query', trackVolatileFunctionAsQuery);
   });
 };