+++ console/src/dataSources/services/postgresql/types.ts
@@ -3,6 +3,7 @@ export type PGFunction = {
   function_schema: string;
   function_definition: string;
   return_type_type: string;
+  function_type: string;
 };
 
 export interface PostgresTable {